requirements.txt
# ============================================================================
# OpenFabrik - Minimal Requirements
# ============================================================================
# Install: pip install -r requirements.txt
#
# For CUDA 12.6 (default). Change cu126 to cu124, cu121, etc. for other versions.
# For CPU-only, remove the --extra-index-url line and torch will install CPU builds.
# ============================================================================

--extra-index-url https://download.pytorch.org/whl/cu126

# ---------------------------------------------------------------------------
# Core: PyTorch (CUDA version resolved automatically via the index URL above)
# ---------------------------------------------------------------------------
torch>=2.0.0
torchvision>=0.15.0

# ---------------------------------------------------------------------------
# Core: Image processing & utilities
# ---------------------------------------------------------------------------
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
PyYAML>=6.0.0
tqdm>=4.66.0
requests>=2.31.0
psutil>=5.9.0

# ---------------------------------------------------------------------------
# LLM prompt generation (Ollama)
# ---------------------------------------------------------------------------
ollama>=0.1.0

# ---------------------------------------------------------------------------
# Diffusion models (FLUX, Zero123++)
# ---------------------------------------------------------------------------
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
huggingface_hub>=0.20.0

# ---------------------------------------------------------------------------
# Object detection & annotation (Grounded-SAM2 pipeline)
# ---------------------------------------------------------------------------
supervision>=0.16.0
pycocotools>=2.0.0

# ---------------------------------------------------------------------------
# YOLO training & inference
# ---------------------------------------------------------------------------
ultralytics>=8.0.0

# ============================================================================
# Optional dependencies - uncomment as needed
# ============================================================================

# --- ONNX export (utilities/yolo_scripts/yolo_export_onnx.py) ---
# onnx>=1.14.0
# onnxruntime-gpu>=1.16.0  # or onnxruntime for CPU-only

# --- TensorRT export (utilities/yolo_scripts/yolo_export_tensorrt.py) ---
# tensorrt  # Install via NVIDIA's instructions for your CUDA version

# --- Benchmarking & visualization (utilities/yolo_scripts/yolo_benchmarking.py) ---
# matplotlib>=3.7.0
# seaborn>=0.12.0
# pandas>=2.0.0
# plotly>=5.15.0
# dash>=2.14.0

# --- ROS2 integration (utilities/ros2_scripts/) ---
# Install via ROS2 apt packages, not pip:
# rclpy, cv_bridge, sensor_msgs, message_filters, sensor_msgs_py

# --- ROS2 data recorder (utilities/ros2_scripts/ros2_data_recorder.py) ---
# open3d>=0.17.0
# imageio>=2.31.0

# --- DINO-X cloud inference (modules/open_set_models/scripts/folder_dino_x_inference.py) ---
# dds-cloudapi-sdk>=0.5.0

# ============================================================================
# External repositories - install separately
# ============================================================================
# These must be cloned and installed in editable mode or added to PYTHONPATH:
#
# Grounded-SAM-2 (scene generation pipeline):
#   git clone https://github.com/IDEA-Research/Grounded-SAM-2
#   export PYTHONPATH=$PYTHONPATH:<path>/Grounded-SAM-2:<path>/Grounded-SAM-2/sam2
#
# Personalize-SAM (reference object pipeline):
#   git clone https://github.com/ZrrSkywalker/Personalize-SAM
#   export PYTHONPATH=$PYTHONPATH:<path>/Personalize-SAM/
#
# ComfyUI (optional, for Kontext augmentation):
#   See https://comfyui-wiki.com/en/install/install-comfyui/install-comfyui-on-linux
